stage 2.3.5

+++ src/main/java/searchengine/services/implementation/IndexingServiceImpl.java
@@ -2,11 +2,10 @@ package searchengine.services.implementation;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import searchengine.config.SitesList;
 import searchengine.customExeptions.PageIndexingException;
 import searchengine.customExeptions.RestartIndexingException;
 import searchengine.customExeptions.StopIndexingException;
-import searchengine.dto.indexing.IndexingResponse;
+import searchengine.dto.indexing.*;
 import searchengine.model.*;
 import searchengine.repositories.IndexRepository;
 import searchengine.repositories.LemmaRepository;
@@ -14,13 +13,11 @@ import searchengine.repositories.PageRepository;
 import searchengine.repositories.SiteRepository;
 import searchengine.services.intarfaces.IndexingService;
 
-import java.io.IOException;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ForkJoinPool;
 
@@ -42,9 +39,7 @@ public class IndexingServiceImpl implements IndexingService {
             }
 
 //            SitesList sitesList = new SitesList();
-//            System.out.println("we are here");
 //            List<searchengine.config.Site> sites = new ArrayList<>(sitesList.getSites());
-//            System.out.println(sites.size());
             List<searchengine.config.Site> sites = new ArrayList<>();
 
             searchengine.config.Site site2 = new searchengine.config.Site();
@@ -75,8 +70,9 @@ public class IndexingServiceImpl implements IndexingService {
                 siteRepository.save(site);
                 site.setId(siteRepository.findByUrl(site.getUrl()).get().getId());
 
-                siteLinks = new SiteLinks(site.getUrl(), site, pageRepository, siteRepository);
-                siteLinks.compute();
+                siteLinks = new SiteLinks(site.getUrl(), site, pageRepository, siteRepository, lemmaRepository, indexRepository);
+                siteLinks.fork();
+                siteLinks.join();
                 site.setStatus(IndexingStatus.INDEXED);
                 site.setId(siteRepository.findByUrl(site.getUrl()).get().getId());
                 siteRepository.save(site);
@@ -90,6 +86,7 @@ public class IndexingServiceImpl implements IndexingService {
             site.setLastError(e.getMessage());
             site.setId(siteRepository.findByUrl(site.getUrl()).get().getId());
             siteRepository.save(site);
+
             return new IndexingResponse(false, e.getMessage());
         }
     }
@@ -122,13 +119,6 @@ public class IndexingServiceImpl implements IndexingService {
     @Override
     public IndexingResponse getIndexingPageResponse(String pageUrl) {
         String url = (URLDecoder.decode(pageUrl, StandardCharsets.UTF_8)).substring(4);
-        LemmasMaker lemmasMaker;
-
-        try{
-            lemmasMaker = LemmasMaker.getInstance();
-        } catch (IOException e) {
-            return new IndexingResponse(false, e.getMessage());
-        }
         Site site;
         String siteUrl = PageUrlConverter.getSiteUrlFromPageUrl(url);
         Optional<Site> siteOptional = siteRepository.findByUrl(siteUrl);
@@ -141,47 +131,20 @@ public class IndexingServiceImpl implements IndexingService {
         } catch (PageIndexingException e) {
             return new IndexingResponse(false, e.getMessage());
         }
+
         String html = HtmlCodeExtractor.getHtmlCode(url);
         String path = url.substring(siteUrl.length());
         Long pageId = pageRepository.findFirstByPathAndSite(path, site).get().getId();
+        if (pageRepository.findById(pageId).get().getCode() != 200) {
+            return new IndexingResponse(false, "Данная страница была некорректно проиндексирована");
+        }
         pageRepository.deleteById(pageId);
         SavePage savePage = new SavePage(pageRepository);
         Page page = savePage.savePageToDatabase(path, html, site);
+        PageIndexing pageIndexing = new PageIndexing(lemmaRepository, indexRepository);
+        String error = pageIndexing.indexingPage(site, page, html);
 
-        Map<String, Integer> lemmasMap = lemmasMaker.collectLemmas(html);
-        for (Map.Entry<String, Integer> entry : lemmasMap.entrySet()) {
-            String key = entry.getKey();
-            Integer value = entry.getValue();
-            Lemma lemma = saveLemma(site, key);
-            saveIndex(page, lemma, value);
-        }
-        return new IndexingResponse(true);
+        return error.isEmpty() ? new IndexingResponse(true) : new IndexingResponse(false, error);
     }
 
-    private Lemma saveLemma(Site site, String key) {
-        Optional<Lemma> lemmaOptional = lemmaRepository.findByLemma(key);
-        Lemma lemma;
-        if (lemmaOptional.isPresent()) {
-            lemma = lemmaOptional.get();
-            Integer frequency = lemma.getFrequency() + 1;
-            lemma.setFrequency(frequency);
-            lemmaRepository.save(lemma);
-        } else {
-            lemma = new Lemma();
-            lemma.setSite(site);
-            lemma.setLemma(key);
-            lemma.setFrequency(1);
-            lemmaRepository.save(lemma);
-        }
-        return lemma;
-    }
-
-    private void saveIndex(Page page, Lemma lemma, Integer count) {
-        Index index = new Index();
-        index.setLemma(lemma);
-        index.setPage(page);
-        Float rank = (float) count / lemma.getFrequency();
-        index.setRank(rank);
-        indexRepository.save(index);
-    }
 }
